:recycle: apply the recase option to table name
