+++ drift_dev/lib/src/analysis/resolver/discover.dart
@@ -2,7 +2,6 @@ import 'package:analyzer/dart/ast/ast.dart' as dart;
 import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/visitor.dart';
-import 'package:recase/recase.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:sqlparser/sqlparser.dart' hide AnalysisError;
 
@@ -217,7 +216,8 @@ class _FindDartElements extends RecursiveElementVisitor<void> {
   }
 
   String _defaultNameForTableOrView(ClassElement definingElement) {
-    return ReCase(definingElement.name).snakeCase;
+    return _discoverStep._driver.options.caseFromDartToSql
+        .apply(definingElement.name);
   }
 
   DartObject? _driftViewAnnotation(ClassElement view) {
